+++ src/components/ControlBar/ConfirmDeleteDialog.js
