+++ src/components/ControlBar/DashboardItemChip.js
