+++ src/components/ControlBar/ShowMoreButton.js
