+++ src/components/ControlBar/__tests__/ClearButton.spec.js
