+++ src/components/ControlBar/__tests__/ConfirmDeleteDialog.spec.js
