+++ src/components/ControlBar/__tests__/DashboardItemChip.spec.js
