+++ src/components/ControlBar/__tests__/DashboardsBar.spec.js
