+++ src/components/ControlBar/__tests__/EditBar.spec.js
