+++ src/components/ControlBar/__tests__/Filter.spec.js
