+++ src/components/ControlBar/__tests__/FilterField.spec.js
