+++ src/components/ControlBar/__tests__/ShowMoreButton.spec.js
