+++ src/components/ControlBar/__tests__/controlBarDimensions.spec.js
