+++ src/components/ControlBar/controlBarDimensions.js
